move hosting under agent
